Added Matlab file via upload
Stochastic solver for the Wigner Fokker Planck (WFP) equation, written in Matlab.
Numerical method: Monte-Carlo (Arianno, Machiavelli), equivalent to the Euler-Maruyama method for stochastic differential equations (the connection between diffusion processes and random walks, as pointed out in the Feynman-Kack formulation). 
The potential is chosen harmonic, because if the Initial Condition is a coherent state, even with diffusive noise the state remains coherent (even if a different one). 
Initial Condition: Harmonic Groundstate (simplest coherent state). Harmonic problem has analytic steady state solution, a coherent state against which we can verify if the numerical solver is getting it right.
Given the problem with a specific diffusion matrix D and friction coefficient gamma, 
the only parameters that can be varied in the numerical method are:
1. N: number of samples
2. T: total evolution time
3. Time-Evolution algorithm: Euler...? Runge-Kutta-2/3/4?
4. Time-step parameter: dt